Added demo script and minor addition to figures.m
`figuresdemo` is a demo script to show the various syntaxes of
`figures`. `figures` itself has received a minor enhancement to support
all the syntax possibilities of the built-in `figure` function: `figure`
now is obsoleted by `figures`.

The readme has been updated accordingly.

`.gitignore` is now updated to ignore MATLAB's autosaves.